requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pypdf
python-multipart
python-dotenv
feedparser
